Move global menu to container
